# kavrigo-platform/apps/web/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kavrigo Control Plane",
    "description": "\nKavrigo control plane.\n\nKavrigo is a platform for building, testing and governing AI trading agents. Agents produce\nstructured, evidence-backed decisions; a deterministic risk engine outside the model decides\nwhether an order intent is permitted.\n\n**This release is paper and research only.** Live execution is disabled and gated on the\nreadiness checklist in the master specification. Backtest and paper results are simulations, not\nlive results, and nothing here is investment advice.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/workspaces/{workspace_id}/paper/policy-bundles": {
      "post": {
        "tags": [
          "paper-policies"
        ],
        "summary": "Store an immutable, unapproved paper-policy candidate pair",
        "operationId": "create_paper_policy_bundle_v1_workspaces__workspace_id__paper_policy_bundles_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ws_[0-9a-f]{32}$",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperPolicyBundleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperPolicyBundleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "paper-policies"
        ],
        "summary": "List paper-policy candidates",
        "operationId": "list_paper_policy_bundles_v1_workspaces__workspace_id__paper_policy_bundles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ws_[0-9a-f]{32}$",
              "title": "Workspace Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PaperPolicyBundleResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/paper/policy-bundles/{bundle_id}": {
      "get": {
        "tags": [
          "paper-policies"
        ],
        "summary": "Read a paper-policy candidate",
        "operationId": "get_paper_policy_bundle_v1_workspaces__workspace_id__paper_policy_bundles__bundle_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^pb_[0-9a-f]{32}$",
              "title": "Bundle Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ws_[0-9a-f]{32}$",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperPolicyBundleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/paper/policy-bundles/{bundle_id}/review": {
      "post": {
        "tags": [
          "paper-policies"
        ],
        "summary": "Record a second-person policy review recommendation without activation",
        "operationId": "review_paper_policy_bundle_v1_workspaces__workspace_id__paper_policy_bundles__bundle_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^pb_[0-9a-f]{32}$",
              "title": "Bundle Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ws_[0-9a-f]{32}$",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperPolicyReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperPolicyReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "paper-policies"
        ],
        "summary": "Read the non-activating review recommendation for a policy candidate",
        "operationId": "get_paper_policy_review_v1_workspaces__workspace_id__paper_policy_bundles__bundle_id__review_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^pb_[0-9a-f]{32}$",
              "title": "Bundle Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ws_[0-9a-f]{32}$",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperPolicyReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/paper/policy-bundles/{bundle_id}/approval": {
      "post": {
        "tags": [
          "paper-policies"
        ],
        "summary": "Approve the exact reviewed paper policies without activating execution",
        "operationId": "approve_paper_policy_bundle_v1_workspaces__workspace_id__paper_policy_bundles__bundle_id__approval_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^pb_[0-9a-f]{32}$",
              "title": "Bundle Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ws_[0-9a-f]{32}$",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaperPolicyApprovalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperPolicyApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "paper-policies"
        ],
        "summary": "Read the non-activating approval for a paper-policy candidate",
        "operationId": "get_paper_policy_approval_v1_workspaces__workspace_id__paper_policy_bundles__bundle_id__approval_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^pb_[0-9a-f]{32}$",
              "title": "Bundle Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^ws_[0-9a-f]{32}$",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperPolicyApprovalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_PaperPolicyBundleResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PaperPolicyBundleResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[PaperPolicyBundleResponse]"
      },
...
      "PaperPolicyApprovalCreate": {
        "properties": {
          "review_id": {
            "type": "string",
            "pattern": "^pr_[0-9a-f]{32}$",
            "title": "Review Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 10,
            "title": "Reason"
          },
          "risk_hash": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Risk Hash"
          },
          "execution_hash": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Execution Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "review_id",
          "reason",
          "risk_hash",
          "execution_hash"
        ],
        "title": "PaperPolicyApprovalCreate",
        "description": "Approve the exact policy documents recorded by an advancing review."
      },
...
      "PaperPolicyBundleCreate": {
        "properties": {
          "limits": {
            "$ref": "#/components/schemas/RiskLimits-Input"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessPolicy"
          },
          "event_risk": {
            "$ref": "#/components/schemas/EventRiskPolicy-Input",
            "default": {
              "block_new_positions_before_macro_minutes": 10,
              "block_new_positions_after_macro_minutes": 5,
              "reduce_size_pct_during_event_risk": "0"
            }
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionCandidate"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "limits",
          "freshness",
          "execution",
          "reason"
        ],
        "title": "PaperPolicyBundleCreate"
      },
      "PaperPolicyBundleResponse": {
        "properties": {
          "bundle_id": {
            "type": "string",
            "pattern": "^pb_[0-9a-f]{32}$",
            "title": "Bundle Id"
          },
          "workspace_id": {
            "type": "string",
            "pattern": "^ws_[0-9a-f]{32}$",
            "title": "Workspace Id"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskPolicy"
          },
          "execution": {
            "$ref": "#/components/schemas/RiskExecutionPolicy"
          },
          "risk_hash": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Risk Hash"
          },
          "execution_hash": {
            "type": "string",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Execution Hash"
          },
          "approval_status": {
            "type": "string",
            "enum": [
              "unapproved",
              "reviewed",
              "changes_requested",
              "approved"
            ],
            "title": "Approval Status"
          },
          "execution_enabled": {
            "type": "boolean",
            "const": false,
            "title": "Execution Enabled",
            "default": false
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bundle_id",
          "workspace_id",
          "risk",
          "execution",
          "risk_hash",
          "execution_hash",
          "approval_status",
          "reason",
          "created_by",
          "created_at"
        ],
        "title": "PaperPolicyBundleResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Identity provider session token",
        "scheme": "bearer"
      }
    }
  }
}